Restructure again: strip down
